ISSUE #4828 - refactor so there are separate 2D and 3D viewer components with thier toolbars
