+++ frontend/src/v4/routes/viewer3D/index.ts
@@ -15,4 +15,4 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { default as ViewerCanvas } from './viewerCanvas.container';
+export { default as Viewer3D } from './viewer3D.container';
