+++ frontend/src/v4/routes/viewer3D/viewer3D.styles.ts
@@ -39,11 +39,12 @@ const getAdditionalStyles = ({ presentationMode, isPresentationPaused }) => {
 	}
 };
 
-export const Container = styled.div<{ visible?: boolean }>`
-	position: absolute;
-	width: 100%;
-	height: 100%;
+export const ViewerContainer = styled.div<{ visible?: boolean }>`
 	visibility: ${({ visible }) => visible ? 'visible' : 'hidden'};
+	height: calc(100vh - 62px);
+	width: 100%;
+	position: relative;
+	box-sizing: border-box;
 `;
 
 export const Border = styled.div<IContainer>`
