+++ frontend/src/v5/ui/components/viewer/drawingViewer/viewer2D.styles.ts
@@ -26,10 +26,3 @@ export const SvgContainer = styled.div`
 	height:100%;
 	overflow: hidden;
 `;
-
-export const DrawingViewerContainer = styled.div`
-	overflow: hidden;
-	width: 100%;
-	height: 100%;
-	user-select: none;
-`;
